Commit 3ba4fb651:
    commit 3d2723effeb6b2d102642896b152396edc78a6a1
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Mon Jan 13 18:05:27 2025 +0700

    Remove path from usage

    commit 966aa74941b9957949e091ad29e43b2a5da8f0bc
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Mon Jan 13 16:30:18 2025 +0700

    End grub config with new line

    commit 4fcc2f24c73c90560b905717a1c86cabd6252154
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Mon Jan 13 14:16:11 2025 +0700

    Remove grub-pc configuration in postinst

    commit 1c405a008742d9332b5a09939ac741b8616ea562
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Mon Jan 13 13:51:56 2025 +0700

    Bump version to 7.6.0

    commit 7a4d7e427a4ca1f89eb5cdd77367ebb7f95c4504
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Mon Jan 13 13:51:23 2025 +0700

    Update package description

    commit 8ddb826bd6f1a0e086fa949da307c40e9e83de8b
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 11 12:17:16 2025 +0700

    Install cron file earlier in postinst

    commit 210e3ca4914c0f94a4ab5e2f7d80662ee7d276ec
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 11 02:16:41 2025 +0700

    Update ucarp path if config file exists

    commit 41d64aa31c526d7fe2fe258b067bdb224a025fa7
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 11 01:06:26 2025 +0700

    Add missing semicolon

    commit a209491157a8786c02368ad37396d828ee909b18
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 10 11:42:13 2025 +0700

    Set nftlb stopped status as successful

    commit d70254d2348bbc09c9c3b993a71d0d94fde7582b
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 10 15:23:43 2025 +0700

    Improve expression

    commit ca9badf8b4a0a97359463c3955195814625f3b40
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 10 15:22:36 2025 +0700

    Add .gitignore file

    commit 0c26c0e2a8b8b7f7707072e6409086f06c5ee988
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 10 14:55:37 2025 +0700

    Add build-deb.sh

    commit 03bf687ab6d5220cbac074c9b270f1e733e5ef98
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 10 14:50:40 2025 +0700

    Set LICENSE to CE root directory as file instead of symlink

    commit fa72dfa282b3d7067667e02a375a419d7087d22b
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Thu Jan 9 19:29:50 2025 +0700

    Move relianoid-ce-cluster.service to /etc for consistency with relianoid.service

    commit 0f4bd5d2b941b8e21c124e2cca43c8fa2d18b8ad
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Thu Jan 9 19:29:50 2025 +0700

    README.md rewrite

    commit 3b188f391756f43b18dcbb8c1216921f0177d2c3
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Jan 8 17:19:54 2025 +0700

    Typo. Fix comment

    commit 84de122e5276b4d25bef2ce4bcbe4936b9fc75a1
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Jan 8 11:58:48 2025 +0700

    Remove unused dependency

    commit 06c5f0c90fc506f8902b62319eda574715e8460d
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 4 15:09:13 2025 +0700

    Remove repetition and improve declaration of command

    commit 77e10559a7749b4d87646de204908a72c946ce3b
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 4 15:07:55 2025 +0700

    Rename variable to errno

    commit a98deb1c3353226e6e6b422cca1df1bfc4d9f9b1
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Sat Jan 4 15:06:02 2025 +0700

    Stop loading relianoid profile when not using bash

    commit 657fe126a3f757b1f436d6392ff8892826561af3
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 3 16:55:59 2025 +0700

    Improvements of prompt behaviour

    commit d53093c679469edb7ecc2505de978fdaccfc6322
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Jan 3 15:42:43 2025 +0700

    Do not include the cluster configuration file as part of the package

    commit 1fe6f722ffa77537f7b03b43f278dcd161e0337e
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Thu Jan 2 19:33:05 2025 +0700

    Fix sed expression

    commit 6f51dbbd8c4e839adf9af530504eacde2a108013
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Thu Jan 2 19:05:07 2025 +0700

    Migrate ucarp config, cleanup files

    commit 43dfe9c3a41c82015dc57b048be3a8fdc71b2654
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Thu Jan 2 16:56:00 2025 +0700

    Add ucarp dependency

    commit 7517dc845353e86c7e55a9ef45445269ffd28c99
    Author: nevola <laura@relianoid.com>
    Date: Fri Dec 27 09:51:15 2024 +0100

    fix undefined subroutine setAPTProxy


    commit 417801323774b9050c94e8a7ffd825e37488b8c8
    Author: nevola <laura@relianoid.com>
    Date: Fri Dec 27 09:44:15 2024 +0100

    fix ssl with passphrase user message


    commit f37825d0e6a0aee019e605224c4ab66d6cf167f3
    Author: nevola <laura@relianoid.com>
    Date: Mon Dec 23 16:48:52 2024 +0100

    avoid to upload ssl certificates with passphrase


    commit e10ff20acda572a58b8e0bfeabad3e2e574455a9
    Author: nevola <laura@relianoid.com>
    Date: Mon Dec 23 14:12:00 2024 +0100

    enable relianoid binaries in the path


    commit 05236c2c89d03186e75fcc50a781f270972553ca
    Author: nevola <laura@relianoid.com>
    Date: Thu Dec 19 16:27:16 2024 +0100

    add support of system http proxy


    commit 46f0fb1d0160303d8f550ca79f365c4ef1fbe3e9
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Dec 11 16:37:46 2024 +0700

    Add runtime information log message and avoid die on error.

    commit d988a6e3294b211a5132506428ae42f4d1773070
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Dec 11 16:36:25 2024 +0700

    Only allow defined graph types

    commit c05ea6ae16a5286c2aea615297c0e8179d79eabe
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Dec 11 16:35:46 2024 +0700

    Sort graph list by name

    commit cfb5d834f6f193853f1b1c72444ed564a8e006e5
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Wed Dec 11 16:34:10 2024 +0700

    Get interface list sorted by name

    commit b6b4d33ab24acb29400305a3a68ddad301f27d46
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Tue Dec 10 16:25:37 2024 +0700

    Remove gpg files from CE files

    commit d8645448775ed6b494b4e05302508ce5636dffeb
    Author: nevola <laura@relianoid.com>
    Date: Thu Dec 5 09:41:53 2024 +0100

    add symlink checkupgrades and checkupdates to noid-updater


    commit 771f07d78fe68d07fb683fabef1dec83cb538db1
    Author: nevola <laura@relianoid.com>
    Date: Wed Dec 4 09:21:43 2024 +0100

    rename upgrading scripts


    commit 622bc7aa1342b32724e2a0bfcafd0c95d390f70e
    Author: nevola <laura@relianoid.com>
    Date: Tue Dec 3 16:34:06 2024 +0100

    fusion checkupgrades and checkupdates in noid-updater


    commit b09dc132f48d20f6e3d62a6e62454408a442a5fa
    Author: nevola <laura@relianoid.com>
    Date: Mon Dec 2 22:06:49 2024 +0100

    fix negative month value in certs


    commit 9f9064380cb677b514693a3d7fef444439533016
    Author: nevola <laura@relianoid.com>
    Date: Mon Dec 2 15:18:33 2024 +0100

    backup: rename restore instead of apply


    commit 60bd072d569436e78655d9034c782c147de36281
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Tue Nov 19 11:56:49 2024 +0700

    Refactor delFGFarm

    commit d413eb73a37f0e9671ba482f29e3699449947fba
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Nov 15 16:57:03 2024 +0700

    relianoid.gpg is not necessary in trusted.gpg.d/

    commit c06a806c2a220f4b16bba6a531f83bdd2c872c23
    Author: Jose P. Garcia <jose@relianoid.com>
    Date: Fri Nov 15 16:54:27 2024 +0700

    keyrings/relianoid.gpg is moved to relianoid-archive-keyring

date: 18:31:30_14/01/25

+++ usr/share/perl5/Relianoid/Farm/Action.pm
@@ -117,6 +117,13 @@ sub _runFarmStart ($farm_name, $writeconf = 0) {
             args   => [ $farm_name, $writeconf ],
         );
     }
+    elsif ($farm_type eq "eproxy" && $eload) {
+        $status = &eload(
+            module => 'Relianoid::EE::Farm::Eproxy::Action',
+            func   => '_runEproxyFarmStart',
+            args   => [ { 'farm_name' => $farm_name, 'write' => $writeconf } ],
+        );
+    }
 
     &setFarmNoRestart($farm_name);
 
@@ -270,6 +277,13 @@ sub _runFarmStop ($farm_name, $writeconf = 0) {
             args   => [ $farm_name, $writeconf ],
         );
     }
+    elsif ($farm_type eq "eproxy" && $eload) {
+        $status = &eload(
+            module => 'Relianoid::EE::Farm::Eproxy::Action',
+            func   => '_runEproxyFarmStop',
+            args   => [ { 'farm_name' => $farm_name, 'write' => $writeconf } ],
+        );
+    }
 
     &setFarmNoRestart($farm_name);
 
@@ -367,10 +381,18 @@ sub runFarmDelete ($farm_name) {
         &runL4FarmDelete($farm_name);
     }
 
-    unlink glob("$configdir/$farm_name\_*\.cfg");
+    if ($farm_type eq "eproxy" && $eload) {
+        $status = &eload(
+            module => 'Relianoid::EE::Farm::Eproxy::Factory',
+            func   => 'runEproxyFarmDelete',
+            args   => [{ farm_name => $farm_name }],
+        );
+    } else {
+        unlink glob("$configdir/$farm_name\_*\.cfg");
 
-    if (!-f "$configdir/$farm_name\_*\.cfg") {
-        $status = 0;
+        if (!-f "$configdir/$farm_name\_*\.cfg") {
+            $status = 0;
+        }
     }
 
     require Relianoid::RRD;
@@ -399,19 +421,27 @@ Returns:
 sub runFarmReload ($farm_name) {
     require Relianoid::Farm::Action;
 
-    if (&getFarmRestartStatus($farm_name)) {
-        &log_info("'Reload' on $farm_name is not executed. 'Restart' is needed.", "FARMS");
-        return 2;
-    }
+    my $farm_type = &getFarmType($farm_name);
     my $status = 0;
 
-    &log_info("running 'Reload' for $farm_name", "FARMS");
+    if ($farm_type eq "http" || $farm_type eq "https") {
+        if (&getFarmRestartStatus($farm_name)) {
+            &log_info("'Reload' on $farm_name is not executed. 'Restart' is needed.", "FARMS");
+            return 2;
+        }
 
-    $status = &_runFarmReload($farm_name);
+        &log_info("running 'Reload' for $farm_name", "FARMS");
 
-    # Reload Farm status from its cfg file
-    require Relianoid::Farm::HTTP::Backend;
-    &setHTTPFarmBackendStatusFromFile($farm_name);
+        $status = &_runFarmReload($farm_name);
+
+        # Reload Farm status from its cfg file
+        require Relianoid::Farm::HTTP::Backend;
+        &setHTTPFarmBackendStatusFromFile($farm_name);
+    }
+    elsif ($farm_type eq "eproxy") {
+        &log_info("running 'Reload' for $farm_name", "FARMS");
+        $status = &_runFarmReload($farm_name);
+    }
 
     return $status;
 }
@@ -438,11 +468,24 @@ sub _runFarmReload ($farm) {
     require Relianoid::Farm::Base;
     return 0 if (&getFarmStatus($farm) ne 'up');
 
-    require Relianoid::Farm::HTTP::Config;
-    my $proxy_ctl = &getGlobalConfiguration('proxyctl');
-    my $socket    = &getHTTPFarmSocket($farm);
+    my $farm_type = &getFarmType($farm);
 
-    $err = &logAndRun("$proxy_ctl -c $socket -R 0");
+    if ($farm_type eq "http" || $farm_type eq "https") {
+        require Relianoid::Farm::HTTP::Config;
+        my $proxy_ctl = &getGlobalConfiguration('proxyctl');
+        my $socket    = &getHTTPFarmSocket($farm);
+
+        $err = &logAndRun("$proxy_ctl -c $socket -R 0");
+    }
+    elsif ($farm_type eq "eproxy" && $eload) {
+        $err = &eload(
+            module => 'Relianoid::EE::Farm::Eproxy::Action',
+            func   => 'runEproxyFarmReload',
+            args   => [{ farm_name => $farm }],
+        );
+        require Relianoid::EE::Cluster;
+        &runClusterRemoteManager('farm', 'reload', $farm);
+    }
 
     return $err;
 }
